+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/KotlinExtensionTest.java
@@ -15,6 +15,9 @@
  */
 package com.diffplug.gradle.spotless;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.io.File;
 import java.io.IOException;
 
 import org.junit.jupiter.api.Test;
@@ -79,6 +82,27 @@ class KotlinExtensionTest extends GradleIntegrationHarness {
 		setFile("src/main/kotlin/Main.kt").toResource("kotlin/ktlint/experimentalEditorConfigOverride.dirty");
 		gradleRunner().withArguments("spotlessApply").build();
 		assertFile("src/main/kotlin/Main.kt").sameAsResource("kotlin/ktlint/experimentalEditorConfigOverride.clean");
+	}
+
+	@Test
+	void testWithInvalidEditorConfigFile() throws IOException {
+		String invalidPath = "invalid/path/to/.editorconfig".replace('/', File.separatorChar);
+
+		setFile("build.gradle").toLines(
+				"plugins {",
+				"    id 'org.jetbrains.kotlin.jvm' version '1.5.31'",
+				"    id 'com.diffplug.spotless'",
+				"}",
+				"repositories { mavenCentral() }",
+				"spotless {",
+				"    kotlin {",
+				"        ktlint().setEditorConfigPath('" + invalidPath.replace("\\", "\\\\") + "')",
+				"    }",
+				"}");
+		setFile("src/main/kotlin/Main.kt").toResource("kotlin/ktlint/experimentalEditorConfigOverride.dirty");
+		String buildOutput = gradleRunner().withArguments("spotlessApply").buildAndFail().getOutput();
+		assertThat(buildOutput).contains("EditorConfig file does not exist: ");
+		assertThat(buildOutput).contains(invalidPath);
 	}
 
 	@Test