+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/json/Json.java
@@ -50,6 +50,11 @@ public class Json extends FormatterFactory {
 		addStepFactory(jackson);
 	}
 
+	public void addBiome(BiomeJson biome) {
+		addStepFactory(biome);
+	}
+
+	@Deprecated
 	public void addRome(RomeJson rome) {
 		addStepFactory(rome);
 	}