+++ testlib/src/main/resources/biome/js/longLineAfter80.js
@@ -1,0 +1,13 @@
+const x = [
+	"Hello",
+	"World",
+	"How",
+	"Are",
+	"You",
+	"Doing",
+	"Today",
+	"Such",
+	"A",
+	"Wondrous",
+	"Sunshine",
+];